Fix problems after rebase
