+++ Civi/Funding/FundingCase/Api4/ActionHandler/RemoteGetFieldsActionHandler.php
@@ -20,7 +20,6 @@ declare(strict_types = 1);
 namespace Civi\Funding\FundingCase\Api4\ActionHandler;
 
 use Civi\Api4\FundingCase;
-use Civi\Api4\Generic\Result;
 use Civi\Funding\Api4\ActionHandler\AbstractRemoteFundingGetFieldsActionHandler;
 use Civi\RemoteTools\Api4\Action\AbstractRemoteGetFieldsAction;
 use CRM_Funding_ExtensionUtil as E;
@@ -29,7 +28,7 @@ final class RemoteGetFieldsActionHandler extends AbstractRemoteFundingGetFieldsA
 
   public const ENTITY_NAME = 'RemoteFundingCase';
 
-  public function getFields(AbstractRemoteGetFieldsAction $action): Result {
+  public function getFields(AbstractRemoteGetFieldsAction $action): array {
     $fields = parent::getFields($action);
     $fields[] = [
       'nullable' => FALSE,
