--- Civi/Funding/PayoutProcess/Api4/ActionHandler/RemoteFundingDrawdown/RemoteGetFieldsActionHandler.php
@@ -1,33 +1,0 @@
-<?php
-/*
- * Copyright (C) 2024 SYSTOPIA GmbH
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as published by
- *  the Free Software Foundation in version 3.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-declare(strict_types = 1);
-
-namespace Civi\Funding\PayoutProcess\Api4\ActionHandler\RemoteFundingDrawdown;
-
-use Civi\Api4\FundingDrawdown;
-use Civi\Funding\Api4\ActionHandler\AbstractRemoteFundingGetFieldsActionHandler;
-
-final class RemoteGetFieldsActionHandler extends AbstractRemoteFundingGetFieldsActionHandler {
-
-  public const ENTITY_NAME = 'RemoteFundingDrawdown';
-
-  protected function getEntityName(): string {
-    return FundingDrawdown::getEntityName();
-  }
-
-}
